refactor(compare): extract C10+C12 to checks/cross_variable.py (Phase A 11/N)
Move check_c10_cross_variable, check_c12_value_mapping_coverage, and the
_C10_SIMPLE_RE regex constant into the checks subpackage. C12 lives here
with C10 (per the recon's family grouping) because both check cross-variable
or cross-mapping consistency rather than per-variable preservation.

cross_variable.py imports `re` (for _C10_SIMPLE_RE), normalize_category_key
from hv_dataqc_common, and four crosswalk helpers (_canonical_phv_id,
_distribution_count_map, _is_null_sentinel_code, _normalize_code) used by
C12's coded-value coverage logic.

This completes Phase A's check-family extractions. All check functions now
live in hv_dataqc/compare/checks/.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 931 → 710 lines.
